Quase funcional

+++ model.py
@@ -1,8 +1,10 @@
+import os
+
 import torch
 import torch.nn as nn
-import torch.optim as optim
 import torch.nn.functional as F
-import os
+import torch.optim as optim
+
 
 class LinearNet(nn.Module):
     def __init__(self, input_size, hidden_size, output_size):
@@ -15,12 +17,13 @@ class LinearNet(nn.Module):
         x = self.linear2(x)
         return x
 
-    def save(self, file_name = 'model.pth'):
+    def save(self, file_name='model.pth'):
         model_folder_path = './model'
         if not os.path.exists(model_folder_path):
             os.makedirs(model_folder_path)
         file_name = os.path.join(model_folder_path, file_name)
         torch.save(self.state_dict(), file_name)
+
 
 class QTrainer:
     def __init__(self, model, lr, gamma):
@@ -33,7 +36,7 @@ class QTrainer:
     def train_step(self, state, action, reward, next_state, done):
         state = torch.tensor(state, dtype=torch.float)
         next_state = torch.tensor(next_state, dtype=torch.float)
-        action = torch.tensor(action, dtype=torch.long)
+        action = torch.tensor(action, dtype=torch.float)
         reward = torch.tensor(reward, dtype=torch.float)
 
         if len(state.shape) == 1:
@@ -41,7 +44,7 @@ class QTrainer:
             next_state = torch.unsqueeze(next_state, 0)
             action = torch.unsqueeze(action, 0)
             reward = torch.unsqueeze(reward, 0)
-            done = (done, )
+            done = (done,)
 
         pred = self.model(state)
